Refactor misc away
